split out types/builder.rs, add module doc comment
